own decorator file, isort fixes
