+++ model.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
-from typing import Optional, TYPE_CHECKING
+
+from typing import TYPE_CHECKING, Optional
 
 from pydantic import BaseModel
 from sqlalchemy import Boolean, Float, Integer, String, update
